+++ Samples~/Scripts/SubUI/SubUIB.cs
@@ -1,7 +1,11 @@
 using Feif.UIFramework;
-using System.Threading.Tasks;
 using UnityEngine;
 using UnityEngine.UI;
+#if USING_UNITASK
+using Task = Cysharp.Threading.Tasks.UniTask;
+#else
+using Task = System.Threading.Tasks.Task;
+#endif
 
 namespace Feif.UI
 {